feat: load schemas for mindset
